backend/requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
requests
sqlalchemy
python-dotenv
